+++ mayan/apps/document_parsing/api_views.py
@@ -5,9 +5,13 @@ from rest_framework.response import Response
 from mayan.apps.documents.models import Document
 from mayan.apps.rest_api import generics
 
-from .models import DocumentPageContent
-from .permissions import permission_content_view
-from .serializers import DocumentPageContentSerializer
+from .models import DocumentPageContent, DocumentTypeSettings
+from .permissions import (
+    permission_content_view, permission_document_type_parsing_setup
+)
+from .serializers import (
+    DocumentPageContentSerializer, DocumentTypeParsingSettingsSerializer
+)
 
 
 class APIDocumentPageContentView(generics.RetrieveAPIView):
@@ -42,3 +46,20 @@ class APIDocumentPageContentView(generics.RetrieveAPIView):
 
         serializer = self.get_serializer(content)
         return Response(serializer.data)
+
+
+class APIDocumentTypeParsingSettingsView(generics.RetrieveUpdateAPIView):
+    """
+    get: Return the document type parsing settings.
+    patch: Set the document type parsing settings.
+    put: Set the document type parsing settings.
+    """
+    lookup_field = 'document_type__pk'
+    lookup_url_kwarg = 'pk'
+    mayan_object_permissions = {
+        'GET': (permission_document_type_parsing_setup,),
+        'PATCH': (permission_document_type_parsing_setup,),
+        'PUT': (permission_document_type_parsing_setup,)
+    }
+    queryset = DocumentTypeSettings.objects.all()
+    serializer_class = DocumentTypeParsingSettingsSerializer