+++ HISTORY.rst
@@ -4,6 +4,7 @@
   GitLab issue #850. Thanks to Olaf (@oohlaf) for the report.
 - Expose the document type OCR settings model via the REST API. Closes
   GitLab issue #851. Thanks to Mike Mansell (@diamondq) for the report.
+- Expose the document type parsing settings model via the REST API.
 
 3.4.12 (2020-07-28)
 ===================